Break out GameView into own file
